backend/requirements.txt
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
tqdm>=4.66.0
urllib3>=2.0.0
supabase>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openpyxl>=3.1.0

# FastAPI and API dependencies
fastapi>=0.115.0
pyjwt>=2.8.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9

# Redis and job queue
redis>=5.0.0
rq>=1.15.0

# OpenAI for AI reports
openai>=1.12.0

# HTTP clients for enrichment
aiohttp>=3.9.0
httpx>=0.27.0

# Web scraping (for future enrichment)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Embeddings
tiktoken>=0.7.0

# Vector database for RAG
chromadb>=1.3.0
